Filling in missing IDs
configuration.HeaderId and configuration.Id
Fixed IDs in test files.
Moved WebClient/Model/ConfigurationHeader class to its own file.
