+++ src/Config.Admin.WebClient/Config.Admin.WebClient/Pages/Applications.razor
