+++ src/Config.Admin.WebClient/Config.Admin.WebClient/wwwroot/appsettings.json
@@ -1,6 +1,6 @@
 {
   "ConnectionStrings": {
-    "AdminApi": "https://localhost:44350/",
+    "AdminApi": "http://localhost:34246/",
     "Api": "http://localhost:5146" 
   }
 }
