+++ src/Config.Middleware/Config.Middleware.TestApi/TestDatabaseFiles/configurations/3f2be342-caa8-4de6-a6fe-aa754cf20ee5.txt
@@ -1,1 +1,1 @@
-{"Id":"3f2be342-caa8-4de6-a6fe-aa754cf20ee5","Name":"RabbitMQ-Login","CreatedUtc":"2022-02-14T20:02:42.647Z","UpdateUtc":"2022-02-14T20:01:27.286Z","Deleted":false,"IsActive":true,"Configurations":[{"HeaderId":"","Id":"","Json":"{\n  \"Username\":\"Goofy\",\n  \"Password\":\"EpicHeroPassword\"\n}","CreatedUtc":"2022-02-14T20:01:36.899Z","IsActive":true,"Applications":"[\"b0876c9d-c46f-4da1-9b92-5f8d80eddbdd\"]","Environments":"[\"5868233c-1d5d-42a6-a93c-9043edabf146\",\"5213b39e-9b17-4d91-bd4b-e54aacddbb49\",\"0f3d9e7a-3979-4130-bef6-b46ca267f565\"]","Deleted":false,"History":[]},{"HeaderId":"","Id":"","Json":"{\n  \"Username\":\"Scrooge\",\n  \"Password\":\"LuckyCoin\"\n}","CreatedUtc":"2022-02-14T20:01:58.972Z","IsActive":true,"Applications":"[\"f2dc0de9-03f8-4c33-9f25-1175c3286fad\"]","Environments":"[\"5868233c-1d5d-42a6-a93c-9043edabf146\",\"5213b39e-9b17-4d91-bd4b-e54aacddbb49\",\"0f3d9e7a-3979-4130-bef6-b46ca267f565\"]","Deleted":false,"History":[]}]}
+{"Id":"3f2be342-caa8-4de6-a6fe-aa754cf20ee5","Name":"RabbitMQ-Login","CreatedUtc":"2022-03-09T22:18:07.818Z","UpdateUtc":"2022-02-14T20:01:27.286Z","Deleted":false,"IsActive":true,"Configurations":[{"HeaderId":"3f2be342-caa8-4de6-a6fe-aa754cf20ee5","Id":"3d993057-498b-462d-9463-7b3ff7e02d48","Json":"{\n  \"Username\":\"Goofy\",\n  \"Password\":\"EpicHeroPassword\"\n}","CreatedUtc":"2022-02-14T20:01:36.899Z","IsActive":true,"Applications":"[\"b0876c9d-c46f-4da1-9b92-5f8d80eddbdd\"]","Environments":"[\"5868233c-1d5d-42a6-a93c-9043edabf146\",\"5213b39e-9b17-4d91-bd4b-e54aacddbb49\",\"0f3d9e7a-3979-4130-bef6-b46ca267f565\"]","Deleted":false,"History":[]},{"HeaderId":"3f2be342-caa8-4de6-a6fe-aa754cf20ee5","Id":"e8676573-9b14-4456-bf43-613aa1daa768","Json":"{\n  \"Username\":\"Scrooge\",\n  \"Password\":\"LuckyCoin\"\n}","CreatedUtc":"2022-02-14T20:01:58.972Z","IsActive":true,"Applications":"[\"f2dc0de9-03f8-4c33-9f25-1175c3286fad\"]","Environments":"[\"5868233c-1d5d-42a6-a93c-9043edabf146\",\"5213b39e-9b17-4d91-bd4b-e54aacddbb49\",\"0f3d9e7a-3979-4130-bef6-b46ca267f565\"]","Deleted":false,"History":[]}]}
