+++ src/Config.Middleware/Config.Middleware.TestApi/TestDatabaseFiles/configurations/9b955216-dfdb-404b-afc8-fcce65209195.txt
@@ -1,1 +1,1 @@
-{"Id":"9b955216-dfdb-404b-afc8-fcce65209195","Name":"RabbitMQ","CreatedUtc":"2022-02-14T20:01:22.042Z","UpdateUtc":"2022-02-14T20:00:27.609Z","Deleted":false,"IsActive":true,"Configurations":[{"HeaderId":"","Id":"","Json":"{\n  \"Server\": \"MyMqServer\",\n  \"Port\": 5672,\n  \"Username\": \"$ref:RabbitMQ-Login#Username\",\n  \"Password\": \"$ref:RabbitMQ-Login#Password\"\n}","CreatedUtc":"2022-02-14T20:01:03.344Z","IsActive":true,"Applications":"[\"b0876c9d-c46f-4da1-9b92-5f8d80eddbdd\",\"f2dc0de9-03f8-4c33-9f25-1175c3286fad\"]","Environments":"[\"5868233c-1d5d-42a6-a93c-9043edabf146\",\"5213b39e-9b17-4d91-bd4b-e54aacddbb49\",\"0f3d9e7a-3979-4130-bef6-b46ca267f565\"]","Deleted":false,"History":[]}]}
+{"Id":"9b955216-dfdb-404b-afc8-fcce65209195","Name":"RabbitMQ","CreatedUtc":"2022-03-09T22:17:45.281Z","UpdateUtc":"2022-02-14T20:00:27.609Z","Deleted":false,"IsActive":true,"Configurations":[{"HeaderId":"9b955216-dfdb-404b-afc8-fcce65209195","Id":"9e4b5800-0a2a-4eda-abaa-ae0890d7f704","Json":"{\n  \"Server\": \"MyMqServer\",\n  \"Port\": 5672,\n  \"Username\": \"$ref:RabbitMQ-Login#Username\",\n  \"Password\": \"$ref:RabbitMQ-Login#Password\"\n}","CreatedUtc":"2022-02-14T20:01:03.344Z","IsActive":true,"Applications":"[\"b0876c9d-c46f-4da1-9b92-5f8d80eddbdd\",\"f2dc0de9-03f8-4c33-9f25-1175c3286fad\"]","Environments":"[\"5868233c-1d5d-42a6-a93c-9043edabf146\",\"5213b39e-9b17-4d91-bd4b-e54aacddbb49\",\"0f3d9e7a-3979-4130-bef6-b46ca267f565\"]","Deleted":false,"History":[]}]}
