+++ src/Config.Middleware/Config.Middleware.TestApi/TestDatabaseFiles/configurations/cde5428a-07b5-43ab-a0d1-8f58fa47b0d8.txt
@@ -1,1 +1,1 @@
-{"Id":"cde5428a-07b5-43ab-a0d1-8f58fa47b0d8","Name":"ConnectionStrings","CreatedUtc":"2022-02-14T20:04:57.348Z","UpdateUtc":"2022-02-14T20:03:15.887Z","Deleted":false,"IsActive":true,"Configurations":[{"HeaderId":"","Id":"","Json":"{\n  \"Default\":\"Data Source=dbserver;Initial Catalog=myDbTest1;User Id=sa;Password=SuperNinjaPassword\",\n  \"AnotherConnection\":\"Server=dbserver;Database=yourDb;Trusted_Connection=True\"\n}","CreatedUtc":"2022-02-14T20:03:22.572Z","IsActive":true,"Applications":"[\"b0876c9d-c46f-4da1-9b92-5f8d80eddbdd\",\"f2dc0de9-03f8-4c33-9f25-1175c3286fad\"]","Environments":"[\"5213b39e-9b17-4d91-bd4b-e54aacddbb49\"]","Deleted":false,"History":[]},{"HeaderId":"","Id":"","Json":"{\n  \"Default\":\"Data Source=dbserver;Initial Catalog=myDbTest2;User Id=sa;Password=SuperNinjaPassword\",\n  \"AnotherConnection\":\"Server=dbserver;Database=yourDb;Trusted_Connection=True\"\n}","CreatedUtc":"2022-02-14T20:03:51.836Z","IsActive":true,"Applications":"[\"b0876c9d-c46f-4da1-9b92-5f8d80eddbdd\",\"f2dc0de9-03f8-4c33-9f25-1175c3286fad\"]","Environments":"[\"0f3d9e7a-3979-4130-bef6-b46ca267f565\"]","Deleted":false,"History":[]},{"HeaderId":"","Id":"","Json":"{\n  \"Default\":\"Data Source=.;Initial Catalog=myDb;Trusted_Connection=True\",\n  \"AnotherConnection\":\"Server=dbserver;Database=yourDb;Trusted_Connection=True\"\n}","CreatedUtc":"2022-02-14T20:04:19.477Z","IsActive":true,"Applications":"[\"b0876c9d-c46f-4da1-9b92-5f8d80eddbdd\",\"f2dc0de9-03f8-4c33-9f25-1175c3286fad\"]","Environments":"[\"5868233c-1d5d-42a6-a93c-9043edabf146\"]","Deleted":false,"History":[]}]}
+{"Id":"cde5428a-07b5-43ab-a0d1-8f58fa47b0d8","Name":"ConnectionStrings","CreatedUtc":"2022-03-09T22:20:00.049Z","UpdateUtc":"2022-02-14T20:03:15.887Z","Deleted":false,"IsActive":true,"Configurations":[{"HeaderId":"cde5428a-07b5-43ab-a0d1-8f58fa47b0d8","Id":"04896439-3640-4c13-8264-77d8c1c3636e","Json":"{\n  \"Default\":\"Data Source=dbserver;Initial Catalog=myDbTest1;User Id=sa;Password=SuperNinjaPassword\",\n  \"AnotherConnection\":\"Server=dbserver;Database=yourDb;Trusted_Connection=True\"\n}","CreatedUtc":"2022-02-14T20:03:22.572Z","IsActive":true,"Applications":"[\"b0876c9d-c46f-4da1-9b92-5f8d80eddbdd\",\"f2dc0de9-03f8-4c33-9f25-1175c3286fad\"]","Environments":"[\"5213b39e-9b17-4d91-bd4b-e54aacddbb49\"]","Deleted":false,"History":[]},{"HeaderId":"cde5428a-07b5-43ab-a0d1-8f58fa47b0d8","Id":"ec9dafc0-18e5-4138-9598-cb288b0def42","Json":"{\n  \"Default\":\"Data Source=dbserver;Initial Catalog=myDbTest2;User Id=sa;Password=SuperNinjaPassword\",\n  \"AnotherConnection\":\"Server=dbserver;Database=yourDb;Trusted_Connection=True\"\n}","CreatedUtc":"2022-02-14T20:03:51.836Z","IsActive":true,"Applications":"[\"b0876c9d-c46f-4da1-9b92-5f8d80eddbdd\",\"f2dc0de9-03f8-4c33-9f25-1175c3286fad\"]","Environments":"[\"0f3d9e7a-3979-4130-bef6-b46ca267f565\"]","Deleted":false,"History":[]},{"HeaderId":"cde5428a-07b5-43ab-a0d1-8f58fa47b0d8","Id":"45ae10a2-c5ed-4263-b18a-76dc382cb1ab","Json":"{\n  \"Default\":\"Data Source=.;Initial Catalog=myDb;Trusted_Connection=True\",\n  \"AnotherConnection\":\"Server=dbserver;Database=yourDb;Trusted_Connection=True\"\n}","CreatedUtc":"2022-02-14T20:04:19.477Z","IsActive":true,"Applications":"[\"b0876c9d-c46f-4da1-9b92-5f8d80eddbdd\",\"f2dc0de9-03f8-4c33-9f25-1175c3286fad\"]","Environments":"[\"5868233c-1d5d-42a6-a93c-9043edabf146\"]","Deleted":false,"History":[]}]}
