+++ src/Config.Middleware/Config.Middleware.TestApi/TestDatabaseFiles/configurations/f5eef302-de4e-4327-93bf-47931eb14d82.txt
@@ -1,1 +1,1 @@
-{"Id":"f5eef302-de4e-4327-93bf-47931eb14d82","Name":"base-configuration","CreatedUtc":"2022-02-14T20:06:40.194Z","UpdateUtc":"2022-02-14T19:57:06.513Z","Deleted":false,"IsActive":true,"Configurations":[{"HeaderId":"","Id":"","Json":"{\n  \"ConnectionStrings\": {\n    \"Default\": \"$ref:ConnectionStrings#Default\"\n  },\n  \"RabbitMQ\": \"$ref:RabbitMQ#\"\n}","CreatedUtc":"2022-02-14T19:59:15.437Z","IsActive":true,"Applications":"[\"b0876c9d-c46f-4da1-9b92-5f8d80eddbdd\",\"f2dc0de9-03f8-4c33-9f25-1175c3286fad\"]","Environments":"[\"5868233c-1d5d-42a6-a93c-9043edabf146\",\"5213b39e-9b17-4d91-bd4b-e54aacddbb49\",\"0f3d9e7a-3979-4130-bef6-b46ca267f565\"]","Deleted":false,"History":[]}]}
+{"Id":"f5eef302-de4e-4327-93bf-47931eb14d82","Name":"base-configuration","CreatedUtc":"2022-03-09T22:18:48.722Z","UpdateUtc":"2022-02-14T19:57:06.513Z","Deleted":false,"IsActive":true,"Configurations":[{"HeaderId":"f5eef302-de4e-4327-93bf-47931eb14d82","Id":"017b697a-642e-447e-b810-cfb017c9f099","Json":"{\n  \"ConnectionStrings\": {\n    \"Default\": \"$ref:ConnectionStrings#Default\"\n  },\n  \"RabbitMQ\": \"$ref:RabbitMQ#\"\n}","CreatedUtc":"2022-02-14T19:59:15.437Z","IsActive":true,"Applications":"[\"b0876c9d-c46f-4da1-9b92-5f8d80eddbdd\",\"f2dc0de9-03f8-4c33-9f25-1175c3286fad\"]","Environments":"[\"5868233c-1d5d-42a6-a93c-9043edabf146\",\"5213b39e-9b17-4d91-bd4b-e54aacddbb49\",\"0f3d9e7a-3979-4130-bef6-b46ca267f565\"]","Deleted":false,"History":[]}]}
